Avoid possible issue of blsdir set to a different path than ostree default
Since the blscfg module will also now be used by ostree-based distros there
is a possible corner case in which a user set the blsdir variable to a BLS
directory path that is different than the default used by ostree.

So to avoid possible issues, only drop the marker file to specify that the
bootloader has support to parse BLS files if this variable hasn't been set.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>